Spawning magic item
This is a magic item that can be used to spawn anything you want. As of
now, only one spell exists that can spawn something, which spawns
barrels. Quite worthwhile, if I do say so myself.

Eventually this will be a rare spawn that will let you cast spells to
create emergency exits from a dungeon.